Refactor: useDeleteTodo
- hooks 폴더로 이동
